refactor(filid): split config-loader into loaders + utils organs (T1-C)
Split core/infra/config-loader/config-loader.ts (431 lines, 12 functions,
2 distinct loader concerns) into an organ-first structure:

- loaders/filid-config.ts       — loadConfig, writeConfig, createDefaultConfig,
                                   loadRuleOverrides, resolveLanguage, initProject
- loaders/rule-docs-manifest.ts — loadRuleDocsManifest, resolveRuleDocSelection,
                                   getRuleDocsStatus, syncRuleDocs
- utils/resolve-git-root.ts     — private helper used by filid-config
- utils/resolve-plugin-root.ts  — private helper used by rule-docs-manifest
- config-loader.ts              — slim facade (30 lines, explicit type re-exports)

rule-docs-manifest -> filid-config is a single-direction sibling import
(loadConfig dependency); no circular deps. Documented in DETAIL.md.

- src/index.ts / src/core/index.ts: empty diff (facade path unchanged)
- config-loader.ts: 431 -> 30 lines
- INTENT.md: 31 lines (<=50)
